connect resizeEffect in split axis
axis-charts.js : use zoomedDomainEffect, allocatedWidth.
chartTypesEffect -> chartsArray, draw().  split out (singular) axis-chart, with each chartsArray in hbs

chart1.js:
Adopt setupFrame() as a method of AxisCharts; ditto  setupChart() and drawChart() -> Chart1.
setupChart() : split overlap() out, and move call to controls() out.
in-axis.js : make available : scaleChanged, domainChanged, zoomedDomain.
split out to form data-types.js :
 from axis-charts.js: featureCountDataProperties, dataConfigs,
 from chart1: featureLocation(), middle(), scaleMaybeInterval(), hoverTextFn(), DataConfig, name2Location(), datum2LocationWithBlock(), datum2Value(),  parsedData, blockData,  blockDataConfig().
frontend/app/ :
 35b5dee 13330 Feb  5 06:34  components/axis-2d.js
 d4fcd13   607 Feb  5 12:08  components/axis-chart.js
 3025f5a 11149 Feb  6 18:02  components/axis-charts.js
 f265212  7010 Feb  5 11:23  components/in-axis.js
 9d1b020  1943 Feb  5 06:34  templates/components/axis-2d.hbs
 f123624   524 Feb  5 11:10  templates/components/axis-charts.hbs
 a9c6323  5302 Feb  6 15:11  utils/data-types.js
 dc1d912 30582 Feb  6 14:33  utils/draw/chart1.js
