+++ frontend/app/templates/components/axis-charts.hbs
@@ -2,7 +2,12 @@
   {{#each blocks as |chartBlock|}}
     {{draw/block-view axis=axis axisID=axisID block=chartBlock blocksData=blocksData}}
   {{/each}}
-  {{log 'chartTypesEffect' chartTypesEffect }}
+  {{#each chartsArray as |chart|}}
+    {{axis-chart axis=axis axisID=axisID blocks=blocks blocksData=blocksData axisCharts=this chart=chart }}
+  {{/each}}
+  {{ resizeEffectHere }}
+  {{ zoomedDomainEffect }}
+
 {{else}}
 <div class="Chart">
   {{content-editable
